+++ packages/components/src/Menu/MenuContext.ts
@@ -29,6 +29,8 @@ import { MenuSharedProps } from './MenuItem'
 
 export interface MenuContextProps {
   disabled?: boolean
+  id?: string
+  isHovered?: boolean
   isOpen?: boolean
   setOpen?: (isOpen: boolean) => void
   triggerElement?: HTMLElement | null