+++ packages/components/src/Menu/MenuItem/MenuItemButton.tsx
@@ -40,6 +40,7 @@ export const MenuItemButton = styled.button<MenuItemButtonProps>`
   ${reset}
   ${space}
 
+  border: 0;
   align-items: center;
   background: transparent;
   color: inherit;