+++ packages/components/src/Popover/Popover.tsx
@@ -42,6 +42,7 @@ import {
   useCallbackRef,
   useControlWarn,
   useFocusTrap,
+  useHovered,
   useScrollLock,
 } from '../utils'
 
@@ -103,11 +104,6 @@ export interface UsePopoverProps {
   pin?: boolean
 
   /**
-   * The element which hovering on/off of will show/hide the triggering element
-   */
-  hoverDisclosureRef?: RefObject<HTMLElement>
-
-  /**
    * Optional, for a controlled version of the component
    */
   setOpen?: (open: boolean) => void
@@ -149,6 +145,11 @@ export interface PopoverProps extends UsePopoverProps {
     ref: Ref<any>,
     className?: string
   ) => JSX.Element
+
+  /**
+   * The element which hovering on/off of will show/hide the triggering element
+   */
+  hoverDisclosureRef?: HTMLElement | null | RefObject<HTMLElement>
 }
 
 function useVerticalSpace(
@@ -326,7 +327,6 @@ export function usePopover({
   isOpen: controlledIsOpen = false,
   onClose,
   placement: propsPlacement = 'bottom',
-  hoverDisclosureRef,
   setOpen: controlledSetOpen,
   triggerElement,
   triggerToggle = true,
@@ -379,32 +379,6 @@ export function usePopover({
     setOpen(false)
     onClose && onClose()
   }
-
-  // Logic to track the hover state of the hoverDisclosureRef and toggle the disclosure
-  const [isHovered, setIsHovered] = useState(hoverDisclosureRef === undefined)
-
-  function handleMouseEnter() {
-    setIsHovered(true)
-  }
-  function handleMouseLeave() {
-    setIsHovered(false)
-  }
-
-  useEffect(() => {
-    const refCurrent = hoverDisclosureRef
-      ? hoverDisclosureRef.current
-      : undefined
-    if (refCurrent) {
-      refCurrent.addEventListener('mouseleave', handleMouseLeave)
-      refCurrent.addEventListener('mouseenter', handleMouseEnter)
-    }
-    return () => {
-      if (refCurrent) {
-        refCurrent.removeEventListener('mouseleave', handleMouseLeave)
-        refCurrent.removeEventListener('mouseenter', handleMouseEnter)
-      }
-    }
-  }, [hoverDisclosureRef])
 
   const popover = !openWithoutElem && isOpen && (
     <ModalContext.Provider
@@ -469,13 +443,19 @@ export function usePopover({
     open: handleOpen,
     popover,
     ref: callbackRef,
-    triggerShown: isOpen || isHovered,
   }
 }
 
-export function Popover({ children, ...props }: PopoverProps) {
-  const { popover, open, ref, isOpen, triggerShown } = usePopover(props)
+export function Popover({
+  children,
+  hoverDisclosureRef,
+  ...props
+}: PopoverProps) {
+  const { popover, open, ref, isOpen } = usePopover(props)
   const childrenOutput = children(open, ref, isOpen ? 'active' : '')
+
+  const [isHovered] = useHovered(hoverDisclosureRef)
+  const triggerShown = isHovered || isOpen
 
   return (
     <>