+++ packages/playground/src/index.tsx
@@ -24,13 +24,14 @@ import { GlobalStyle } from '@looker/components'
 import { theme } from '@looker/design-tokens'
 import { ThemeProvider } from 'styled-components'
 
-import { SelectDemo } from './Select/SelectDemo'
-
+import { MenuDemo } from './Menu/MenuDemo'
 const App: React.FC = () => {
   return (
     <ThemeProvider theme={theme}>
-      <GlobalStyle />
-      <SelectDemo />
+      <>
+        <GlobalStyle />
+        <MenuDemo />
+      </>
     </ThemeProvider>
   )
 }